Restructure src folder + solving subsequent path issues
