+++ src/paradigma/gait/gait_analysis_config.py
