+++ src/paradigma/heart_rate/heart_rate_analysis_config.py
