+++ src/paradigma/heart_rate/heart_rate_quality.py
