+++ src/paradigma/tremor_matlab/feat_extraction/ExtractBandSignalsRMS.m
