+++ src/paradigma/tremor_matlab/feat_extraction/PSDExtrAxis.m
